two modes for llms.txt
